requirements.txt
# Runtime dependencies for PiWebcam
# Install with: pip3 install -r requirements.txt

# Camera support (Raspberry Pi only)
picamera>=1.13

# Image processing for motion detection
Pillow>=8.0.0
numpy>=1.19.0
